Revert "Modificacao sem perder criticidade"
This reverts commit a28081f79828e8be310d198d6e3ec3a8803be462.

+++ metric/nli/nli_aligner.py
@@ -17,15 +17,9 @@ class NLIAligner:
         self.tokenizer = AutoTokenizer.from_pretrained(model_name)
 
         self.device = device
-        # self.model = AutoModelForSequenceClassification.from_pretrained(model_name).to(
-        #     device
-        # )
-        #Modificado aqui tbm
-        self.model = AutoModelForSequenceClassification.from_pretrained(
-                model_name,
-                torch_dtype=torch.float16
-            ).to(device)
-
+        self.model = AutoModelForSequenceClassification.from_pretrained(model_name).to(
+            device
+        )
         self.batch_size = batch_size
         self.max_length = max_length
